Add actor movement and terrain elevation layer.

+++ Assets/Sandbox/Tile.cs
@@ -4,14 +4,14 @@ using System.Collections.Generic;
 using UnityEngine;
 
 public class Tile {
-    public int elevation { get; private set; }
+    public float elevation { get; private set; }
     public int temperature { get; private set; }
     public TerrainData terrain { get; private set; }   
     public Vector2Int position { get; private set; }
 
     private Level level; 
 
-    public Tile(Level level, Vector2Int position, TerrainData terrain, int elevation, int temperature)
+    public Tile(Level level, Vector2Int position, TerrainData terrain, float elevation, int temperature)
     {
         this.level = level;
         this.position = position;
